Security, input features, and defect-fix overhaul Reviewed via a 5-dimension adversarial pass; all 9 found issues fixed. Transport security (TLS + client cert): - Replace AllCertsTrustManager (trust-all -> MITM) with SHA-256 fingerprint   TOFU (FingerprintTrustManager). Pin only AFTER the TLS handshake completes,   so a responder that fails the handshake (wrong server/transient/MITM that   can't finish) is never pinned. System-CA fallback now also enforces hostname   verification (gated, so self-signed TOFU still works). - Auto-generate a self-signed RSA-2048 client cert in AndroidKeyStore for   PeerAuth (mTLS); show this device's client fingerprint in Settings. - Fix the TLS handshake (was a non-blocking busy-spin/hang), make OP_WRITE   on-demand (was a 100%-CPU selector busy-loop; interestOps now mutated only   on the selector thread to avoid stranding outbound messages), and handle   BUFFER_OVERFLOW. - Atomic trust-store writes; persist pinned fingerprints in app-private store. Input: - Magnitude-aware mouse wheel (accumulators, horizontal tilt, momentum). - Mouse back button -> system Back (forward has no Android global action);   fix readByte().toUInt() sign-extension on mouse button ids. - Keyboard volume up/down/mute -> AudioManager (Deskflow wire KeyIDs   0xE0AF/0xE0AE/0xE0AD), fired on key-down only. - Configurable pointer-speed slider (delta-scaling, screen-clamped, re-syncs   on screen entry); new AppPrefs.mouse_sensitivity field. Defects: - Fix getState() ANR (runBlocking on a never-completing StateFlow) and move   DataStore writes off the main thread. - Declare CONNECTION_SERVICE_PERMISSION at signature level (was squatting-prone). - Make ConnectionService a foreground service (connectedDevice) + boot start   via startForegroundService. - Client: dispose the connection executor (no leaked thread/zombie reconnect),   run teardown off the main thread; SimpleEventEmitter -> CopyOnWriteArrayList;   MessageHandler.onDispose no longer holds the monitor across awaitTermination. - Protocol hardening: frame-length validation + per-frame try/catch   (CVE-2020-15117) and clipboard size caps (CVE-2026-65976); readString is now   the inverse of writeString; DynamicByteBuffer InputStream contract + pop offset. - Sensitive-clipboard flag on both clipboard paths; network_security_config   (cleartext off); allowBackup off. Build/CI/tests/cleanup: - Hilt removed (was wired but unused); proguard-rules.pro added; versionCode   collision guards; CI verify job + test/lint gate (release.yml, version.yml). - Tests: MessageTemplate parametrized over all templates; DynamicByteBuffer   (10 tests). - Dead-code cleanup; CLI startup race + hostname-resolution guard. Verify: ./gradlew :client:compileKotlin :client:test :client-cli:compileKotlin :app:compileDebugKotlin :app:lintDebug -> BUILD SUCCESSFUL (13 tests pass).

+++ client/src/main/java/org/tfv/deskflow/client/io/msgs/MouseDownMessage.kt
@@ -31,7 +31,9 @@ class MouseDownMessage(
 ) : Message(MessageType.DMOUSEDOWN) {
 
     override fun readData(inStream: DataInputStream, dataSize: Int) {
-        buttonId = inStream.readByte().toUInt()
+        // readByte() is signed; toUByte() treats it as unsigned 0..255 so ids
+        // >= 128 are not sign-extended to 0xFFFFFFxx.
+        buttonId = inStream.readByte().toUByte().toUInt()
     }
 
     override fun writeData(outStream: DataOutputStream) {
